requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
langchain_google_genai
llama-index 
langchain-community
requests
beautifulsoup4
langchain_experimental
qdrant-client
ragas
pandas
data